sub2/backend/requirements.txt
attrs==19.3.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
distributed==1.26.0
Django==2.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-pandas==0.6.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-webpack-loader==0.2.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.14
drf-yasg==1.17.1
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jsonpointer==2.0
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.0.1
konlpy==0.5.2
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.42.0
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oauthlib==3.1.0
packaging==20.3
pandas==0.24.1
pandas-datareader==0.7.0
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.4.1
seaborn==0.9.0
spyder-kernels==0.5.0
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
tensorboard==2.1.1
tensorflow-gpu-estimator==2.1.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.7
